+++ assets/js/app.js
@@ -147,6 +147,10 @@ const listeners = {
   },
   "phx:chat:focus": (e) => { const el = document.querySelector(e.detail.to); setTimeout(() => el.focus(), 100); },
   "phx:chat:change": (e) => { const el = document.querySelector(e.detail.to); el.innerHTML = e.detail.content; },
+  "phx:chat:bulk-change": (e) => {
+    const elements = document.querySelectorAll(e.detail.to);
+    elements.forEach((el) => { el.innerHTML = e.detail.content; });
+  },
   "phx:scroll-to-bottom": (e) => {
     setTimeout(() => {
       const chatContent = document.querySelector('.a-content-block');