+++ CHANGELOG.md
@@ -11,11 +11,14 @@ Here we write upgrading notes for brands. It's a team effort to make them as
 straightforward as possible.
 
 ### Added
+- Accept all invites in dialog
 
 ### Changed
 
 ### Fixed
 - UI: uploads scroll to top on uploader mount 
+- submitting of edited message by keyword
+
 ## 2023-03-22_4082136___2023-03-23_e2c9aa8
 
 ### Added